requirements.txt
# MCP 客户端与 LangChain
langchain-mcp-adapters>=0.1.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-core>=0.3.0

# 示例 MCP 服务
fastmcp>=2.0.0

# 环境变量
python-dotenv>=1.0.0
